Combine add email into manage emails page
