+++ src/account/ManageEmail.tsx
@@ -1,9 +1,9 @@
-import React, { useState, useEffect } from "react";
+import React, { useState, useEffect, SyntheticEvent } from "react";
 import { AppActions } from "../store";
-import { primaryEmail, ensureEmails } from "../store/emails/api";
+import { addEmail, ensureEmails } from "../store/emails/api";
 import Field from "../common/Field";
 import { ArchieState } from "../store/archie/types";
-import { EmailState } from "../store/emails/types";
+import { Email, EmailState } from "../store/emails/types";
 import EmailCard from "./EmailCard";
 
 interface ManageEmailPageProps {
@@ -30,11 +30,53 @@ export const ManageEmail: React.FC<ManageEmailPageProps> = (
     fetchData();
   }, [props.actions, props.emails]);
 
+  function handleAddEmailSubmit(event: SyntheticEvent) {
+    event.preventDefault();
+    let emailObj: Email = {
+      email: email,
+      verified: false,
+      primary: false
+    }
+    addEmail(
+      props.actions,
+      emailObj,
+    ).then(status => {
+      if (status.ok) {
+        setSaved(true);
+        setErrors({});
+      } else {
+        setErrors(status.body);
+      }
+    });
+  }
+
+  let savedConfirmation = saved ? (
+    <div className="notification is-success limited-width">
+      {props.archie.copy.add_email.success}
+    </div>
+  ) : null;
+
   return (
     <section className="section">
       <h1 className="title is-size-1">{props.archie.copy.manage_email.title}</h1>
       <p>{props.archie.copy.manage_email.description}</p>
 
+      <section className="section">
+        {savedConfirmation}
+        <form className="limited-width" onSubmit={handleAddEmailSubmit}>
+          <Field label="New Email Address" errors={errors.email}>
+            <input
+              type="email"
+              className={errors.email ? "input is-danger" : "input"}
+              value={email}
+              onChange={event => setEmail(event.target.value)}
+            />
+          </Field>
+          <button className="button is-link" type="submit">
+            {props.archie.copy.buttons.add_email}
+          </button>
+        </form>
+      </section>
       <div className="columns is-multiline">
         {items.map(item => (
           <div className="column is-4" key={item.email}>
