+++ src/account/routing.tsx
@@ -2,7 +2,6 @@ import React from "react";
 import { AppProps } from "../store";
 import { AuthProps } from "../store/auth/types";
 import { ProtectedRoute } from "../common/routing";
-import { AddEmail } from "./AddEmail";
 import { ManageEmail } from "./ManageEmail";
 import { ChangePassword } from "./ChangePassword";
 import { ProfilePage } from "./ProfilePage";
@@ -13,9 +12,6 @@ export const getRoutes = (props: AppProps) => {
   const routes = [
     <ProtectedRoute exact path="/profile/change-password" {...authProps}>
       <ChangePassword actions={props.actions} archie={props.archie} />
-    </ProtectedRoute>,
-    <ProtectedRoute exact path="/profile/add-email" {...authProps}>
-      <AddEmail actions={props.actions} archie={props.archie} />
     </ProtectedRoute>,
     <ProtectedRoute exact path="/profile/manage-email" {...authProps}>
       <ManageEmail actions={props.actions} archie={props.archie} emails={props.emails} />
