+++ src/store/emails/reducers.ts
@@ -23,20 +23,26 @@ export function emailReducers(
         hydrated: true
       };
 
-      // update email responds only with the email address string, not an object
+      let email = action.email;
+      // // update email responds only with the email address string, not an object
       if (typeof(action.email) === 'string') {
-        let emailAddress = action.email;
-        let email: Email = {
-          email: emailAddress,
-          verified: true, // TODO will this actually be true?
+        email = {
+          email: action.email,
+          verified: true,
           primary: true
         };
-        incomingObject.emails.forEach((item, index) => {
-          // replace matching email data here
-          if (item.email === emailAddress) {
-            incomingObject.emails[index] = email;
-          }
-        });
+      }
+      // incoming email is either an update or a new one
+      let foundMatch = false;
+      incomingObject.emails.forEach((item, index) => {
+        // replace matching email data here
+        if (item.email === action.email.email) {
+          foundMatch = true;
+          incomingObject.emails[index] = action.email;
+        }
+      });
+      if (!foundMatch) {
+        incomingObject.emails.push(action.email)
       }
       return Object.assign({}, state, incomingObject);
     }
